Create DIR_24_CHROMOSOME_LOOP
DIR_24_CHROMOSOME_LOOP.pl – writes temporary perl-scripts, which will be used to carry out the mapping to specific genome ranges. The total of perl scripts resulting from this script equals [24*n*r], where the ‘n’ stands for the number of files split, and ‘r’ for the number of range categories in the analysis. If you want to analyze 100 sets of variants against 16 different range categories, then the number of perl scripts written in this step is 24*100*16 = 38,400.

b. The ‘zNAMES_in_Directory.z’ (see above) is pasted in parenthesis after ‘my @Dir = ‘ in line 60 of DIR_24_CHROMOSOME_LOOP.pl